+++ packages/driver/src/cy/commands/index.ts
@@ -71,7 +71,7 @@ export const allCommands = {
   Misc,
   Popups,
   Navigation,
-  Querying,
+  ...Querying,
   Request,
   Sessions,
   Screenshot,
