+++ packages/driver/src/cy/commands/querying/index.ts
@@ -1,0 +1,7 @@
+import * as Querying from './querying'
+import * as Within from './within'
+
+export {
+  Querying,
+  Within,
+}
