+++ packages/driver/src/cy/commands/querying/within.ts
@@ -1,0 +1,105 @@
+import _ from 'lodash'
+
+import { $Command } from '../../../cypress/command'
+import $errUtils from '../../../cypress/error_utils'
+
+export default (Commands, Cypress, cy, state) => {
+  Commands.addAll({ prevSubject: ['element', 'document'] }, {
+    within (subject, options, fn) {
+      let userOptions = options
+      const ctx = this
+
+      if (_.isUndefined(fn)) {
+        fn = userOptions
+        userOptions = {}
+      }
+
+      options = _.defaults({}, userOptions, { log: true })
+
+      if (options.log) {
+        options._log = Cypress.log({
+          $el: subject,
+          message: '',
+          timeout: options.timeout,
+        })
+      }
+
+      if (!_.isFunction(fn)) {
+        $errUtils.throwErrByPath('within.invalid_argument', { onFail: options._log })
+      }
+
+      // reference the next command after this
+      // within.  when that command runs we'll
+      // know to remove withinSubject
+      const next = state('current').get('next')
+
+      // backup the current withinSubject
+      // this prevents a bug where we null out
+      // withinSubject when there are nested .withins()
+      // we want the inner within to restore the outer
+      // once its done
+      const prevWithinSubject = state('withinSubject')
+
+      state('withinSubject', subject)
+
+      // https://github.com/cypress-io/cypress/pull/8699
+      // An internal command is inserted to create a divider between
+      // commands inside within() callback and commands chained to it.
+      const restoreCmdIndex = state('index') + 1
+
+      cy.queue.insert(restoreCmdIndex, $Command.create({
+        args: [subject],
+        name: 'within-restore',
+        fn: (subject) => subject,
+      }))
+
+      state('index', restoreCmdIndex)
+
+      fn.call(ctx, subject)
+
+      const cleanup = () => cy.removeListener('command:start', setWithinSubject)
+
+      // we need a mechanism to know when we should remove
+      // our withinSubject so we dont accidentally keep it
+      // around after the within callback is done executing
+      // so when each command starts, check to see if this
+      // is the command which references our 'next' and
+      // if so, remove the within subject
+      const setWithinSubject = (obj) => {
+        if (obj !== next) {
+          return
+        }
+
+        // okay so what we're doing here is creating a property
+        // which stores the 'next' command which will reset the
+        // withinSubject.  If two 'within' commands reference the
+        // exact same 'next' command, then this prevents accidentally
+        // resetting withinSubject more than once.  If they point
+        // to differnet 'next's then its okay
+        if (next !== state('nextWithinSubject')) {
+          state('withinSubject', prevWithinSubject || null)
+          state('nextWithinSubject', next)
+        }
+
+        // regardless nuke this listeners
+        cleanup()
+      }
+
+      // if next is defined then we know we'll eventually
+      // unbind these listeners
+      if (next) {
+        cy.on('command:start', setWithinSubject)
+      } else {
+        // remove our listener if we happen to reach the end
+        // event which will finalize cleanup if there was no next obj
+        cy.once('command:queue:before:end', () => {
+          cleanup()
+
+          state('withinSubject', null)
+        })
+      }
+
+      return subject
+    },
+  })
+}
